refactor(context): Remove new Context for flow query params
Because:
* We should only need flowQueryParams in a few components and since using Context isn't free, we can just pass it directly

This commit:
* Refactors out the new Context provider and mocks in favor of passing these values directly into the components that need them
* Moves AppErrorBoundary up the tree
* Moves a couple types out of react-app-env.d.ts

Closes FXA-6163
