--- packages/fxa-settings/src/models/FlowContext.ts
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const FlowContext = createContext<any>(undefined);
